city pick

+++ app/src/main/java/com/example/hrautomation/presentation/model/social/EventFilterParam.kt
@@ -5,6 +5,6 @@ import java.io.Serializable
 data class EventFilterParam(
     val name: String?,
     val date: String?,
-    val cityId: Int?,
+    val cityId: Long?,
     val format: String?
 ) : Serializable
